+++ frontend/openchat-agent/src/services/common/chatMappersV2.ts
@@ -566,6 +566,7 @@ export function message(value: TMessage): Message {
         senderContext: mapOptional(value.sender_context, senderContext),
         ogPreviews: (value.og_previews ?? []).map(ogPreview),
         messagePreviews: [], // this will be rehydrated later
+        moderationFlags: value.moderation_flags,
     };
 }
 
@@ -2233,6 +2234,7 @@ export function communitySummary(value: TCommunityCanisterCommunitySummary): Com
         localUserIndex,
         isInvited: value.is_invited ?? false,
         verified: value.verified ?? false,
+        moderationFlags: value.moderation_flags ?? 0,
     };
 }